+++ Kel16/src/FillingAndTopping.java
@@ -1,3 +1,4 @@
+
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
@@ -9,12 +10,12 @@
  */
 public interface FillingAndTopping{
     //filling dan topping dari roti
-    int keju();
-    int coklat();
-    int krimVanila();
-    int redBean();
-    int sosis();
-    int smokeBeef();
-    int bawangBombay();
+    double keju();
+    double coklat();
+    double krimVanila();
+    double redBean();
+    double sosis();
+    double smokeBeef();
+    double bawangBombay();
     
 }
